backend/requirements.txt
aiofiles==0.6.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
Jinja2==3.0.0rc1
lxml==4.6.3
MarkupSafe==2.0.0rc2
pydantic==1.8.1
PyJWT==2.0.1
requests==2.25.1
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
